// src/utils/schemas/confirmFormDataSchema.ts
import { z } from 'zod'

export const confirmFormSchema = z.object({
  name: z.string().min(3, { message: 'O nome precisa no mínimo 3 caracteres' }),
  email: z.string().email({ message: 'Digite um e-mail válido' }),
  observations: z.string().nullable(),
})

export type ConfirmFormDataType = z.infer<typeof confirmFormSchema>